create LiveData helper function for UseCase withParameter
